+++ src/main/java/webserver/managers/LoginManager.java
@@ -9,6 +9,7 @@ import java.util.UUID;
 import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.Cookie;
 import webserver.request.HttpRequest;
 import webserver.response.HttpResponse;
 import webserver.response.HttpResponseBody;
@@ -16,7 +17,6 @@ import webserver.response.HttpResponseHeader;
 import webserver.response.HttpResponseStatusLine;
 import webserver.response.HttpStatus;
 import webserver.response.HttpVersion;
-import webserver.session.Cookie;
 import webserver.session.Session;
 import webserver.session.SessionStore;
 import webserver.utils.HttpMessageUtils;
