+++ src/test/java/webserver/response/HttpResponseHeaderTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import webserver.session.Cookie;
+import webserver.Cookie;
 
 class HttpResponseHeaderTest {
